+++ lang/en/errors.php
@@ -5,4 +5,8 @@ return [
     'invalid_file_type' => 'The file type is invalid.',
     'file_too_loud' => 'The file is too loud.',
     'upload' => 'An error occured during upload.',
+    'image_too_small_width' => 'The image is too narrow (minimum width: :min px).',
+    'image_too_large_width' => 'The image is too wide (maximum width: :max px).',
+    'image_too_small_height' => 'The image is too short (minimum height: :min px).',
+    'image_too_large_height' => 'The image is too tall (maximum height: :max px).',
 ];