+++ lang/fr/errors.php
@@ -5,4 +5,8 @@ return [
     'invalid_file_type' => 'Le type de fichier est invalide.',
     'file_too_loud' => 'Le fichier est trop lourd.',
     'upload' => 'Une erreur est survenue lors de l’envoi du fichier.',
+    'image_too_small_width' => 'L’image est trop étroite (largeur minimale : :min px).',
+    'image_too_large_width' => 'L’image est trop large (largeur maximale : :max px).',
+    'image_too_small_height' => 'L’image est trop petite (hauteur minimale : :min px).',
+    'image_too_large_height' => 'L’image est trop haute (hauteur maximale : :max px).',
 ];